Consolidate users growth chart component
